requirements.txt
langchain>=0.1.14
langchain-community>=0.0.28
openai>=1.11.0
python-dotenv>=1.0.1
duckduckgo-search>=4.4.0
